Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/manifest-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/manifest.test/manifest.json because it does not appear in the manifest-src directive of the Content Security Policy.
ALERT: Pass
